move config, library_data, load_data file to utils package
